Split merge.py and context.py
